Make canvaskit image reusable for multiple surfaces
