+++ namui/src/namui/skia/web/surface.rs
@@ -1,5 +1,4 @@
 use super::*;
-use wasm_bindgen::JsValue;
 
 pub(crate) struct Surface {
     canvas_kit_surface: CanvasKitSurface,
@@ -20,19 +19,6 @@ impl Surface {
     }
     pub fn canvas(&self) -> &Canvas {
         &self.canvas
-    }
-    pub fn make_image_from_texture_source(
-        &self,
-        src: JsValue, // NOTE: It can also be an HTMLVideoElement or an HTMLCanvasElement.
-        info: Option<PartialImageInfo>,
-        src_is_premul: Option<bool>,
-    ) -> Image {
-        let info = info.map(|info| info.into_js_object());
-        let image = self
-            .canvas_kit_surface
-            .makeImageFromTextureSource(src, info, src_is_premul);
-        let image = image.makeCopyWithDefaultMipmaps();
-        Image::new(image)
     }
 }
